Added more functionality to paragraph node

+++ Assets/Scripts/Graph/Editor/Nodes/ParagraphNode.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using Book;
 using Graph.Editor.Nodes;
 using UnityEditor;
@@ -12,17 +13,16 @@ namespace Nodes
     [Serializable]
     public class ParagraphNode : Node
     {
-        [Input] public Choice inputChoices;
+        [Input] public ChoiceEditor choice;
         [field: SerializeField] public Page PageTemplate { get; private set; }
-        [field: SerializeField, Output(dynamicPortList = true)] public Choice[] Choices { get; private set; }
+        [field: SerializeField, Output(dynamicPortList = true)] public ChoiceEditor[] EditorChoices { get; private set; }
         public Action<Object> OnPageChanged;
         
         [Tooltip("It is created automatically when you select a page template")]
         [field: SerializeField, ReadOnly] public Page PageVariant { get; private set; }
-        private int index;
         
         public override object GetValue(NodePort port) {
-            return port.fieldName == "Choices" ? Choices : null;
+            return port.fieldName == "Choices" ? EditorChoices : null;
         }
         
         public void CreatePrefabVariant(Object prefabTemplate)
@@ -30,8 +30,7 @@ namespace Nodes
             if (prefabTemplate is not Page page) { return; }
 
             var instance = PrefabUtility.InstantiatePrefab(page) as Page;
-            index = Utils.PageCounter.GetAnotherIndex();
-            PageVariant = PrefabUtility.SaveAsPrefabAsset(instance.gameObject, Utils.Consts.PagePrefabsPath + "/" + "Page" + index + ".prefab").GetComponent<Page>();
+            PageVariant = PrefabUtility.SaveAsPrefabAsset(instance.gameObject, Utils.Consts.PagePrefabsPath + "/" + Utils.Consts.PageNamePrefix + Utils.PageCounter.GetAnotherIndex() + ".prefab").GetComponent<Page>();
             
             DestroyImmediate(instance.gameObject); ;
         }
@@ -39,10 +38,61 @@ namespace Nodes
         public void DeletePagePrefab()
         {
             if(PageVariant == null) { return; }
-            AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(PageVariant));
+
+            if (!Utils.PrefabNameUtils.TryGetIndexFromName(PageVariant.name, out int index))
+            {
+                Debug.LogError("Can't delete page prefab, page variant name is invalid");
+                return;
+            }
+                
             Utils.PageCounter.ReturnIndexToPool(index);
+            AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(PageVariant));
         }
-        
+
+        public void CreateNewChoice()
+        {
+            var instantiatePrefab = (Page)PrefabUtility.InstantiatePrefab(PageVariant);
+            var newChoice = Instantiate(PageVariant.ChoiceTemplate, instantiatePrefab.ChoiceParent);
+            
+            instantiatePrefab.Choices.Add(newChoice);
+            
+            PrefabUtility.ApplyPrefabInstance(instantiatePrefab.gameObject, InteractionMode.UserAction);
+            DestroyImmediate(instantiatePrefab.gameObject);
+            UpdateChoices();
+        }
+
+        public void UpdateChoices()
+        {
+            if (PageVariant == null)
+                return;
+            
+            PageVariant.Choices.Clear();
+            PageVariant.OnValidate();
+
+            EditorChoices = PageVariant.Choices.Select(x =>
+            {
+                var choiceEditor = new ChoiceEditor
+                {
+                    choice = x
+                };
+                
+                return choiceEditor;
+            }).ToArray();
+        }
+
+        protected override void Init()
+        {
+            base.Init();
+            if(PageVariant == null)
+                return;
+
+            ChoiceEditorDrawer.OnContentChanged += node =>
+            {
+                UpdateChoices();
+            };
+            UpdateChoices();
+        }
+
         private void OnDestroy()
         {
             DeletePagePrefab();
